Added list of contributors
This is an unfinished commit. There is still much UI work that needs
to be done.

The statistics repository has a new method: `getContributors`. This
method will return top contributors by a set of measurable categories.

The /contributors list will display these in a "Hall of Fame"-like
environment. The deign of this page is yet to be decided. ALso the
order of the items in the list.